+++ internal/provider/resource_catalystcenter_anycast_gateway_test.go
@@ -60,7 +60,6 @@ const testAccCcAnycastGatewayPrerequisitesConfig = `
 resource "catalystcenter_area" "test" {
   name        = "Area1"
   parent_name = "Global"
-  depends_on  = [catalystcenter_ip_pool.test]
 }
 resource "catalystcenter_ip_pool" "test" {
   name             = "MyPool1"
@@ -80,7 +79,6 @@ resource "catalystcenter_fabric_site" "test" {
   site_id                     = catalystcenter_area.test.id
   pub_sub_enabled             = false
   authentication_profile_name = "No Authentication"
-  depends_on = [catalystcenter_area.test]
 }
 resource "catalystcenter_fabric_virtual_network" "test" {
   virtual_network_name = "SDA_VN1"